+++ work/resources/ts/contexts/TotalContext.tsx
@@ -1,10 +1,9 @@
 import React, { createContext } from 'react';
 import { StylesContextProvider } from './childContexts/StylesContext';
 import { AuthContextProvider } from './childContexts/AuthContext';
-import { GetBoardsContextProvider } from './childContexts/GetBoardsContext';
+import { ApiBoardsContextProvider } from './childContexts/ApiBoardsContext';
 import { StoreBoardContextProvider } from './childContexts/StoreBoardContext';
 import { UpdateBoardContextProvider } from './childContexts/UpdateBoardContext';
-import { DeleteBoardContextProvider } from './childContexts/DeleteBoardContext';
 import { ShowCardsContextProvider } from './childContexts/ShowCardsContext';
 import { BoardItemContextProvider } from './childContexts/BoardItemContext';
 
@@ -14,19 +13,17 @@ export const TotalContextProvider: React.FC = props => {
   return (
     <StylesContextProvider>
       <AuthContextProvider>
-        <GetBoardsContextProvider>
+        <ApiBoardsContextProvider>
           <StoreBoardContextProvider>
             <UpdateBoardContextProvider>
-              <DeleteBoardContextProvider>
-                <ShowCardsContextProvider>
-                  <BoardItemContextProvider>
-                    {props.children}
-                  </BoardItemContextProvider>
-                </ShowCardsContextProvider>
-              </DeleteBoardContextProvider>
+              <ShowCardsContextProvider>
+                <BoardItemContextProvider>
+                  {props.children}
+                </BoardItemContextProvider>
+              </ShowCardsContextProvider>
             </UpdateBoardContextProvider>
           </StoreBoardContextProvider>
-        </GetBoardsContextProvider>
+        </ApiBoardsContextProvider>
       </AuthContextProvider>
     </StylesContextProvider>
   );
